--- src/main/java/com/jiro4989/tkfm/util/ValidationUtil.java
@@ -1,18 +1,0 @@
-package com.jiro4989.tkfm.util;
-
-public class ValidationUtil {
-  /** Returns `true` when the string format `value` is integer. */
-  public static boolean isInteger(String value) {
-    if (value.matches("^0\\d+$")) return false;
-    if (value.matches("^\\d+$")) return true;
-    return false;
-  }
-
-  /** Returns `true` when the string format `value` is double. */
-  public static boolean isDouble(String value) {
-    if (value.matches("^0\\.\\d+$")) return true;
-    if (value.matches("^[^0]\\d*\\.\\d+$")) return true;
-    if (value.matches("^[\\d]+$")) return true;
-    return false;
-  }
-}
